+++ src/util/ui.hpp
@@ -48,4 +48,6 @@ namespace util::ui {
     BadgeLayout createLayout(std::string _badgePNG, float _scale, cocos2d::CCPoint _pos, std::string _id, cocos2d::CCNode* _badgeWrapper);
 
     uibuilder::Build<cocos2d::CCSprite> createBadge(BadgeLayout layout);
+
+    uibuilder::Build<cocos2d::CCSprite> createBadgeIfSpecial(cocos2d::ccColor3B color, BadgeLayout layout);
 }